terrasketch/requirements.txt
# Django Core
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1

# Database
psycopg2==2.9.9
django-environ==0.11.2

# Celery
celery==5.3.6
redis==5.0.1
django-celery-beat==2.6.0

# File handling
Pillow==10.3.0
WeasyPrint==61.2
ezdxf==1.3.0

# API integrations
anthropic==0.25.6
stripe==9.5.0
requests==2.31.0

# Authentication & Security
django-otp==1.5.0
python-decouple==3.8

# Development & Testing
pytest-django==4.8.0
factory-boy==3.3.0
black==24.4.0
isort==5.13.2

# Monitoring
sentry-sdk==2.3.0

# Other utilities
python-dateutil==2.9.0
pytz==2024.1